backend/requirements.txt
Flask==3.0.2
gunicorn==21.2.0

SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

Flask-CORS==4.0.0

python-dotenv==1.0.1
